D111: 300+ project improvements — __init, timestamp, _Static_assert, stack buffers, API docs, prefix casing, narrowing fixes
Batch A: __init annotations (70 functions in fs/drivers/kernel/memory)
Batch C: Build timestamp (3) + _Static_assert (12 struct size assumptions)
Batch F: Large stack buffer→kmalloc conversion (23 buffers across 9 files)
Batch H: API kernel-doc (48 functions across 13 zero-doc files)
Batch I: kprintf prefix casing (60+ subsystems to uppercase)
Batch I: size_t/int narrowing fixes (20 comparisons)
Batch J: partial C99 variable placement + initcall ordering (timed out, partial)

~130 files changed
Zero-warning build with -Wall -Wextra -Werror

+++ src/fs/debugfs.c
@@ -99,7 +99,7 @@ static int debugfs_vfs_read(void *priv, const char *path, void *buf,
     if (debugfs_entries[idx].type == 1 && debugfs_entries[idx].read_fn) {
         int len = 0;
         debugfs_entries[idx].read_fn(cbuf, &len);
-        if (len > (int)max_size) len = (int)max_size;
+        if ((size_t)len > max_size) len = (int)max_size;
         if (len < 0) len = 0;
         cbuf[len] = '\0';
         *out_size = (uint32_t)len;
@@ -123,7 +123,7 @@ static int debugfs_vfs_read(void *priv, const char *path, void *buf,
     } else if (debugfs_entries[idx].type == 3 && debugfs_entries[idx].read_fn) {
         int len = 0;
         debugfs_entries[idx].read_fn(cbuf, &len);
-        if (len > (int)max_size) len = (int)max_size;
+        if ((size_t)len > max_size) len = (int)max_size;
         if (len < 0) len = 0;
         cbuf[len] = '\0';
         *out_size = (uint32_t)len;
